stories refactor
