+++ elements/layercontrol/src/enums/index.js
@@ -1,0 +1,13 @@
+export {
+  STORIES_LAYER_S2,
+  STORIES_LAYER_OSM,
+  STORIES_LAYER_REGION,
+  STORIES_LAYER_SENTINEL_HUB,
+  STORIES_MAIN_MAP_LAYERS,
+  STORIES_MAP_STYLE,
+  STORIES_LAYER_EOX_CLOUDLESS_2021,
+  STORIES_LAYER_EOX_CLOUDLESS_2020,
+  STORIES_LAYER_EOX_CLOUDLESS_2019,
+  STORIES_LAYER_DEFORESTED_BIOMASS,
+  STORIES_LAYER_TERRAIN_LIGHT,
+} from "./stories";
