+++ elements/layercontrol/stories/tabs.js
@@ -1,0 +1,13 @@
+import { html } from "lit";
+
+export const Tabs = {
+  render: () => html`
+    <eox-layercontrol-tabs
+      .noShadow=${false}
+      .actions=${["delete"]}
+      .tabs=${["info", "opacity", "config"]}
+    ></eox-layercontrol-tabs>
+  `,
+};
+
+export default Tabs;
